+++ src/Buckle/Silver/Buckle/CodeAnalysis/Symbols/Primitives.cs
@@ -4,7 +4,7 @@ namespace Buckle.CodeAnalysis.Symbols;
 internal sealed class TypeSymbol : Symbol {
     public static readonly TypeSymbol Error = new TypeSymbol("?");
     public static readonly TypeSymbol Int = new TypeSymbol("int");
-    public static readonly TypeSymbol Float = new TypeSymbol("float");
+    public static readonly TypeSymbol Decimal = new TypeSymbol("decimal");
     public static readonly TypeSymbol Bool = new TypeSymbol("bool");
     public static readonly TypeSymbol String = new TypeSymbol("string");
     public static readonly TypeSymbol Any = new TypeSymbol("any");
